85 more, and the position pair the movers write through
Coordinates: every x, y, nx, ny and sx local, and the routines that had
kept a wide one - paddle_row_offsets, field_backdrop, walker_draw,
walker_step. bonus_steer and the four bonus_move_* take a **pointer** to
the pair, so those narrowed too, and verify.c's copies with them.

Positions and offsets: centre, top, pos, next, w, word. Counters and
lengths that are not `for` heads: i, k, r, count, rows, which - and with
them draw_run's and draw_text's count, xor_sprite_16xn's rows and
field_marks_wide's, which are how many of a thing rather than 32-bit
numbers. read_new_key's which and ending_blob's pos.

Checked: bonus_steer, bonus_script, the four movers, bonus_spawn,
bonus_update, walker_step, walker_draw, field_backdrop,
paddle_row_offsets, xor_sprite_16xn, draw_cursor, ending_column,
ending_blob and screen_all_levels_done all agree with the original, and
2,000 frames on panel 8, 1,500 on level 11 and 1,200 through the bonus
are identical throughout.

+++ tools/verify.c
@@ -396,7 +396,7 @@ static int32_t dispatch(uint32_t routine, const uint16_t *r)
          * capsule at the origin, take a different branch, and draw
          * twice from the PRNG where the original drew nothing. The
          * failure that reported was the harness's, not the port's. */
-        uint32_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
+        uint16_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
         g_result = bonus_steer(&entity_ptr(r[R_BX])->p.anim, &x, &y);
         return 1;
     }
@@ -406,7 +406,7 @@ static int32_t dispatch(uint32_t routine, const uint16_t *r)
          * capsule at the origin, take a different branch, and draw
          * twice from the PRNG where the original drew nothing. The
          * failure that reported was the harness's, not the port's. */
-        uint32_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
+        uint16_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
         g_result = bonus_script(&entity_ptr(r[R_BX])->p.anim, &x, &y);
         return 1;
     }
@@ -419,22 +419,22 @@ static int32_t dispatch(uint32_t routine, const uint16_t *r)
      * file. What is compared is the decision, which is the part with the cell
      * arithmetic in it; the step itself is one `inc` behind that decision. */
     case 0x3c66: {                      /* bonus_move_right(bx, cl, al) */
-        uint32_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
+        uint16_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
         g_result = bonus_move_right(&entity_ptr(r[R_BX])->p.anim, &x, &y);
         return 1;
     }
     case 0x3caf: {                      /* bonus_move_up */
-        uint32_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
+        uint16_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
         g_result = bonus_move_up(&entity_ptr(r[R_BX])->p.anim, &x, &y);
         return 1;
     }
     case 0x3cf3: {                      /* bonus_move_left */
-        uint32_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
+        uint16_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
         g_result = bonus_move_left(&entity_ptr(r[R_BX])->p.anim, &x, &y);
         return 1;
     }
     case 0x3d3c: {                      /* bonus_move_down */
-        uint32_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
+        uint16_t x = r[R_CX] & 0xff, y = r[R_AX] & 0xff;
         g_result = bonus_move_down(&entity_ptr(r[R_BX])->p.anim, &x, &y);
         return 1;
     }
